requirements.txt
fastapi
uvicorn
httpx
python-dotenv
PyJWT
beautifulsoup4
pydantic
slowapi
markdownify
groq